# Optimize name-id-lookup for mapblock serialization