--- a/src/mapblock.cpp
+++ b/src/mapblock.cpp
@@ -321,25 +321,20 @@ static void correctBlockNodeIds(const NameIdMapping *nimap, MapNode *nodes,
 	std::unordered_set<content_t> unnamed_contents;
 	std::unordered_set<std::string> unallocatable_contents;
 
-	bool previous_exists = false;
-	content_t previous_local_id = CONTENT_IGNORE;
-	content_t previous_global_id = CONTENT_IGNORE;
+	// Used to cache local to global id lookup.
+	IdIdMapping mapping_cache;
 
 	for (u32 i = 0; i < MapBlock::nodecount; i++) {
 		content_t local_id = nodes[i].getContent();
-		// If previous node local_id was found and same than before, don't lookup maps
-		// apply directly previous resolved id
-		// This permits to massively improve loading performance when nodes are similar
-		// example: default:air, default:stone are massively present
-		if (previous_exists && local_id == previous_local_id) {
-			nodes[i].setContent(previous_global_id);
+
+		if (mapping_cache.get(local_id) != 0xFFFF) {
+			nodes[i].setContent(mapping_cache.get(local_id));
 			continue;
 		}
 
 		std::string name;
 		if (!nimap->getName(local_id, name)) {
 			unnamed_contents.insert(local_id);
-			previous_exists = false;
 			continue;
 		}
 
@@ -348,16 +343,13 @@ static void correctBlockNodeIds(const NameIdMapping *nimap, MapNode *nodes,
 			global_id = gamedef->allocateUnknownNodeId(name);
 			if (global_id == CONTENT_IGNORE) {
 				unallocatable_contents.insert(name);
-				previous_exists = false;
 				continue;
 			}
 		}
 		nodes[i].setContent(global_id);
 
 		// Save previous node local_id & global_id result
-		previous_local_id = local_id;
-		previous_global_id = global_id;
-		previous_exists = true;
+		mapping_cache.set(local_id, global_id);
 	}
 
 	for (const content_t c: unnamed_contents) {